move routing table to a different file
