+++ include/opendht/dht.h
@@ -26,6 +26,7 @@
 #include "utils.h"
 #include "network_engine.h"
 #include "scheduler.h"
+#include "routing_table.h"
 
 #include <string>
 #include <array>
@@ -366,60 +367,6 @@ private:
     private:
         std::list<std::weak_ptr<Node>> cache_4;
         std::list<std::weak_ptr<Node>> cache_6;
-    };
-
-    struct Bucket {
-        Bucket() : cached() {}
-        Bucket(sa_family_t af, const InfoHash& f = {}, time_point t = time_point::min())
-            : af(af), first(f), time(t), cached() {}
-        sa_family_t af {0};
-        InfoHash first {};
-        time_point time {time_point::min()};             /* time of last reply in this bucket */
-        std::list<std::shared_ptr<Node>> nodes {};
-        sockaddr_storage cached;  /* the address of a likely candidate */
-        socklen_t cachedlen {0};
-
-        /** Return a random node in a bucket. */
-        std::shared_ptr<Node> randomNode();
-    };
-
-    class RoutingTable : public std::list<Bucket> {
-    public:
-        using std::list<Bucket>::list;
-
-        InfoHash middle(const RoutingTable::const_iterator&) const;
-
-        std::vector<std::shared_ptr<Node>> findClosestNodes(const InfoHash id, time_point now, size_t count = TARGET_NODES) const;
-
-        RoutingTable::iterator findBucket(const InfoHash& id);
-        RoutingTable::const_iterator findBucket(const InfoHash& id) const;
-
-        /**
-         * Return true if the id is in the bucket's range.
-         */
-        inline bool contains(const RoutingTable::const_iterator& bucket, const InfoHash& id) const {
-            return InfoHash::cmp(bucket->first, id) <= 0
-                && (std::next(bucket) == end() || InfoHash::cmp(id, std::next(bucket)->first) < 0);
-        }
-
-        /**
-         * Return true if the table has no bucket ore one empty buket.
-         */
-        inline bool isEmpty() const {
-            return empty() || (size() == 1 && front().nodes.empty());
-        }
-
-        /**
-         * Return a random id in the bucket's range.
-         */
-        InfoHash randomId(const RoutingTable::const_iterator& bucket) const;
-
-        unsigned depth(const RoutingTable::const_iterator& bucket) const;
-
-        /**
-         * Split a bucket in two equal parts.
-         */
-        bool split(const RoutingTable::iterator& b);
     };
 
     struct SearchNode {
